refactor: move ipc request validation and authZ to ipc class
